+++ app/components/Activity/ActivityIcon.jsx
